make pixel render a component and add icon routes
