+++ web/scenes/Onboarding/Login/layout/index.tsx
@@ -1,7 +1,8 @@
 import { DecoratedButton } from "@/components/DecoratedButton";
-import { WorldIcon } from "@/components/Icons/WorldIcon";
+import { WorldTextLogo } from "@/components/Icons/WorldTextLogo";
 import { SizingWrapper } from "@/components/SizingWrapper";
 import { urls } from "@/lib/urls";
+import Link from "next/link";
 import { ReactNode } from "react";
 
 export const LoginLayout = (props: { children: ReactNode }) => {
@@ -10,7 +11,13 @@ export const LoginLayout = (props: { children: ReactNode }) => {
       <header className="border-b border-grey-100">
         <SizingWrapper>
           <div className="flex items-center justify-between py-2.5">
-            <WorldIcon />
+            <Link
+              href="/"
+              aria-label="World"
+              className="inline-flex items-center"
+            >
+              <WorldTextLogo className="h-7 w-auto" />
+            </Link>
 
             <DecoratedButton
               href={urls.api.authLogin()}
